refactor: onclick 이벤트 myButton 처리에서 page처리로 변경
